# Add ConnectionIdlePingTime setting