This allows clients to opt-out of pinging the server when a connection is retrieved from the pool, if it's been idle less than a certain amount of time. This removes a probably-unnecessary roundtrip to the server on recently-returned connections, which can improve throughput under heavy load.

### Before

Job | Runtime |     Toolchain |     Mean |     Error |    StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |
    Op/s |  Gen 0 | Allocated |
---------- |-------- |-------------- |---------:|----------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|--------:|-------:|----------:|
net47 |     Clr |   CsProjnet47 | 184.6 us | 3.5489 us | 3.3197 us | 0.8571 us | 178.0 us | 181.9 us | 184.9 us | 186.5 us | 191.3 us | 5,416.6 | 1.7090 |  11.84 KB |
netcore20 |    Core | .NET Core 2.0 | 157.1 us | 0.7065 us | 0.6263 us | 0.1674 us | 156.0 us | 156.7 us | 157.1 us | 157.5 us | 158.4 us | 6,363.6 | 1.2207 |   3.13 KB |

### After 
 
Job | Runtime |     Toolchain |      Mean |     Error |    StdDev |    StdErr |       Min |        Q1 |    Median |        Q3 |
 Max |     Op/s |  Gen 0 | Allocated |
---------- |-------- |-------------- |----------:|----------:|----------:|----------:|----------:|----------:|----------:|----------:|----------:|---------:|-------:|----------:|
net47 |     Clr |   CsProjnet47 | 115.20 us | 0.9376 us | 0.8771 us | 0.2265 us | 113.87 us | 114.69 us | 114.85 us | 115.97 us | 116.55 us |  8,680.5 | 1.0986 |    7.6 KB |
netcore20 |    Core | .NET Core 2.0 |  91.13 us | 0.4125 us | 0.3858 us | 0.0996 us |  90.56 us |  90.66 us |  91.21 us |  91.39 us |  91.83 us | 10,973.0 | 0.9766 |   5.03 KB |
 
 CC @sebastienros