@bgrainger 2018-03-20
Added in [0.37.0](https://github.com/mysql-net/MySqlConnector/releases/tag/0.37.0).