@bgrainger 2020-06-11
Yes; it may change or be removed in the future: https://github.com/mysql-net/MySqlConnector/issues/596.

Changes to connection pooling (https://github.com/mysql-net/MySqlConnector/issues/178) may necessitate a change to this setting.