@felixpelletier 2021-03-24
Our team uses the connector for many short connections to a database with a very stable connection. Having ConnectionIdlePingTime come out of the _experimental_ phase would be great for us, as right now we have a use case where we spend about 40% of our time just pinging the server. 

Should I open an issue to see if anyone else would interested in this getting out if the experimental phase?