@bgrainger 2021-03-24
Are you running >= 1.3.0 (with the new background connection reset code)?