@felixpelletier 2021-03-24
It does. Before the update, we spent something like 40-50% of our time resetting connections, so that seems to have helped a lot.(Thank you and everyone involved for that improvement, by the way.).