@bgrainger 2021-03-25
I think my current preference would be to make this a simple binary setting, and recommend using a retry policy (e.g., [Polly](https://github.com/App-vNext/Polly)) around all usages of MySqlConnector, rather than baking in just one retry policy (based on pinging the server).

That said, it does still feel quite strange that `MySqlConnection.Open` may or may not result in an open connection depending on the underlying connection pool:

* no pooled connection available; has to open a new one; might be slow; >99.9% likelihood of being valid
* pooled connection available; returns almost instantly; connection state may be invalid if the server has closed it or something else has gone wrong

OTOH, this may just be an unavoidable side-effect of how connection pooling is typically implemented for ADO.NET and not addressable until a consistent ADO.NET API for pooling is added (if ever).