@bgrainger 2021-03-25
The available options one could choose would be:

* reset connection upon return to pool, no check on retrieval
* reset connection upon return to pool, ping on retrieval (current default)
* reset connection upon retrieval
* no reset; ping on retrieval
* no reset; no ping on retrieval

There are currently three connection string options to control this, and only certain combinations of them are valid; it would be nice to simplify that.