@felixpelletier 2021-03-25
Come to thing of it, wouldn't that scenario be problematic?:

1. The connection pool gets really big with a lot of connections open. Say 50.
2. The application does some unrelated processing for a while.
3. The MySQL server closes all connections (timeout).
4. The app goes back to using the database and opens a connection.
5. The connection is bad, it tries another one, still bad.
6. And so on until it has tried all 50 connections.

In that scenario, if the MySQL server really is dead, it would take 50 calls to Open() to make sure of it.