@bgrainger 2021-03-25
[`ConnectionIdleTimeout`](https://mysqlconnector.net/connection-options/#ConnectionIdleTimeout) should be cleaning up those connections in the background: a) to avoid this very problem, b) to avoid tying up resources on the server unnecessarily; a sudden spike in load shouldn't result in 50 extra connections being kept open forever.

(The default is pretty aggressive, but anything lower than MySQL's default `wait_timeout` value should work well in practice.)