@bgrainger 2021-03-25
> I'm curious why you'd prefer binary, though.

Mostly because it's simpler, and because I _think_ the problem of shutting down idle connections is generally getting handled (in a gracious manner) client-side through `ConnectionIdleTimeout`.