@bgrainger 2021-10-01
I created an issue to change the default behaviour of `Connection Reset=false` to never ping (and just return a pooled connection): https://github.com/mysql-net/MySqlConnector/issues/1042.